pyproject.toml
[project]
name = "amplifier-builder-coach"
version = "0.1.0"
description = "Metacognitive coaching bundle for AI builders"
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pyyaml>=6.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
